Migrate 0.2.1 code, massive refactor to symbol class structure

+++ shardscript-semantics/src/main/kotlin/com/tsikhe/shardscript/semantics/core/Architecture.kt
@@ -1,9 +1,7 @@
 package com.tsikhe.shardscript.semantics.core
 
-import java.math.BigInteger
-
 interface Architecture {
-    val distributedPluginCost: BigInteger
-    val defaultNodeCost: BigInteger
-    val costUpperLimit: BigInteger
+    val distributedPluginCost: Long
+    val defaultNodeCost: Long
+    val costUpperLimit: Long
 }
